requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.1
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
phonenumberslite==8.12.55
Pillow==9.2.0
psycopg2==2.9.4
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
